Added lesson 1 and lesson 2 flash cards.
This flashcards cover the important aspect of the first two units.
They are formatted for blank index cards with the answers being slightly
lighter in color so they can't been seen from looking at the questions.
They also have information in the footer to be easier to organized and
grouped by their respective lesson groups.