Added flashcards for lessons 2 and 3. Created a quiz for lesson 1.
Lesson 2 has one set of flashcards covering both topics (it's a short
section). Lesson 3 has two sets of flash cards, one for topic 1 and
one for topic 2. Both are word files already formated for 3x5 index
cards.

Lesson 1 has a terminal based quiz. All that is needed is python 3.8
or greater installed. Use the run_quiz.bat file to take the quiz.
Users are able to select how many questions they want and a report
is displayed at the end showing which questions were wrong and what
the correct answer was. Plans to make this look better to come,
currently focusing on getting the information together.